feat: add feature_chart test skeleton
I'm still figuring out how to structure integration tests and so I
anticipate the structure continuing to evolve.

This commits demonstrates adding an interceptor to the client that adds
an auth header and tests the feature chart get_chart rpc.

There was some difficulty with the generated code; it appeared that
rust_analyzer could not infer the client's type correctly (or maybe
it was vscode struggling I'm not sure). To work around this I explicitly
set the response type.