feat!: refactor server structure
Quiet a few changes here and given that I'm still exploring how to best
structure a rust-based service I'm going to cheat a bit with the size
and scope of this commit. Hopefully this restructure will help focus
future commits.

Some noteworthy changes:

1. I've separated register, which is a public route, from user functions,
which are private. I've added a basic implementation so that a user
can provide a `uid` which, for the moment, is a sha256. The general
idea is for the app store to compute the sha256 from the machine-id,
username, and potentially some salt. We need more exploration here bc
if its salted then there is no way for the user to recover their
`uid` between installs on the same machine. This is probably fine until
a fully fledged solution lands which includes proper authentication (for
example using SSO).

2. I dropped mongo deps in favour of postgres and started
exploring how it could work inside a rust-service. My current approach
is to pass a Arc wrapped Infrastructure struct down to services
allowing them to access database connections in a thread safe manner. I
haven't come across a DI approach for rust which I think makes sense
given that DI can hide a lot of detail. Although there is some support
for passing shared resources around with hyper Extensions, or in the
case of axum a context object, I again feel this hides detail which
doesn't feel like idomatic rust. Do keep in mind I'm still a rust newb.

3. I've followed group by feature approach where each area generally
follows this pattern:

- interface.rs exposes this features public surface
- infrastructure.rs groups things that this feature drives (ie a db)
- use_case.rs groups the feature's use cases
- errors.rs domain errors

You can learn more about this approach [here](https://herbertograca.com/2017/11/16/explicit-architecture-01-ddd-hexagonal-onion-clean-cqrs-how-i-put-it-all-together/).