feat: add Context middleware layer to support authn
Simple interceptors (aka middleware) have limited access to the request.
This can be problematic if, for example, a subsequent interceptor needs
to see which method the request is destined for.

To work around this, I've created a `hyper::service::Service` to extract
request data and attach it as a request extension. Subsequent services
or middleware can then access this object as required.

In the cause of authentication, the interceptor can check if the path
requested should be protected or not. One final point, in the ideal
setup we would be able to package groups of routes and pick and choose
which interceptors apply. However, in practice this isn't supported by
tonic. In fact, tonic supports a single 'layer' for all routes. Perhaps
in layering will be more flexible.

Signed-off-by: Tim Holmes-Mitra <tim.holmes-mitra@canonical.com>