feat: use lazy initialisation for an INFRA singleton
Attempting to pass an Infrastructure Singleton through the layers was
proving to be quiet clunky. Perhaps I was doing it wrong, but moving
to a once_cell global singleton has made it much easier as now the
app building initializes the infrastructure (currently dbs and jwt) and
then that singleton can be accesses where ever it's needed. The values
inside the Singleton do need to be wrapped in an Arc and potentially
Mutex should the calling scope require mutations.

Another motivation for this was to support passing the Jwt instance
into services. It was causing ownership and borrowing issues in my
Middleware layer which I was unsure how to resolve. I decided to try
this approach and I now feel it lead to a cleaner structure.

Signed-off-by: Tim Holmes-Mitra <tim.holmes-mitra@canonical.com>