feat: add register + authenticate and vote skeleton
Previously there was a single Login method used both to create the
account and to login. Although they serve similar purposes conceptually
they are not the same. So this change requires a registration step
to create the account. This process will issue a JWT valid for 24
hours. Subsequently, the app store can get a refresh token by calling
authenticate.

Because `user_id = sha254(machine-id+username)` if the OS is reinstalled
a new account will be created. In time we will resolve this through some
SSO solution but for now this solution is sufficient.

This also adds a skeleton to start the voting implementation.

Signed-off-by: Tim Holmes-Mitra <tim.holmes-mitra@canonical.com>