feat(vote): implement vote and update user id name
This commit adds functionality for users to cast and update votes. Each
vote is uniquely identified by a tuple (user_id_fk, snap_id,
snap_revision). In the event of a conflict (i.e., a user attempts to
cast a vote on a tuple they've already voted on), the vote is updated.

Additionally, the `user_id` field in the `users` table has been renamed
to `client_hash`. This better represents the field's purpose and value.
From the perspective of the client, it is submitting an id, but from
the service's perspective, this id is treated as a hash that uniquely
identifies the client. The new name, `client_hash`, therefore, is more
descriptive and accurate.

Signed-off-by: Tim Holmes-Mitra <tim.holmes-mitra@canonical.com>