fix(cicd): remove experimental optional field from .protos
In proto2, fields were required by default, and you have to explicitly
 mark fields as optional. If a required field is missing when parsing,
 then it's an error.

proto3 removed the concept of required fields. All fields in proto3 are
optional by default, meaning they can be left unset. It doesn't mean
"optional" in a sense that you can choose whether to include the field
or not in your message, it means that these fields can be missing or can
be present.

There is an 'optional' type in proto3 but its experimental and so this
breaks building in image in cicd. This removes the optional modifier
and handles the empty string case in the interface.

Signed-off-by: Tim Holmes-Mitra <tim.holmes-mitra@canonical.com>