fix: handle get_votes when no records exist
If there are no votes for a snap_id then a default VoteSummary should be
returned.

Ideally we'd be able to validate if the id maps to a real snap, but
the service is not connected to the snap store or a local instance of
snapd so we cannot do that check.