feat: update deps and rust version to 1.75.0
Most debs had either a minor or patch version bumps, however, hyper
was deliberately skipped. hyper 1.0 was released in Dec 2023 with
breaking changes. The issue is that other deps have not caught up
to hyper's 1.0 and so we can't bump them until they do, or we have
time to explore a bigger refactor.