ci: Workflow to check, build and publish oci image
Builds the service as a docker image and upload it to ghcr.io. We want
to use a rock in the future but using a rocks right now will slow our
progress.

I've refined the check jobs to include cargo audit and clippy and
make their output available in github's ui. In time we can evolve this
to trigger another action (like create an issue for a security
vulnerability).